rename abstractions/ to data/, drop /abstraction/ from schema IDs, replace Tab with Website
