+++ package.json
@@ -30,7 +30,7 @@
   "scripts": {
     "lint": "eslint \"**/*.js\"",
     "lint:fix": "eslint \"**/*.js\" --fix",
-    "test": "node --experimental-vm-modules ../../../node_modules/jest/bin/jest.js"
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js"
   },
   "author": "me@idnc.sk",
   "license": "AGPL-3.0-or-later",
