+++ src/schemas/data/Document.js
@@ -1,48 +1,38 @@
 'use strict';
 
-import Document, { documentSchema as baseDocumentSchema } from '../BaseDocument.js';
+import BaseDocument, { documentSchema } from '../BaseDocument.js';
 import { z } from 'zod';
 
-const DOCUMENT_SCHEMA_NAME = 'data/abstraction/todo';
+const DOCUMENT_SCHEMA_NAME = 'data/document';
 const DOCUMENT_SCHEMA_VERSION = '2.0';
 
 const documentDataSchema = z.object({
     schema: z.string(),
     schemaVersion: z.string().optional(),
-    data: z.object({
-        title: z.string(),
-        description: z.string().optional(),
-        completed: z.boolean().optional(),
-        dueDate: z.string().datetime().optional(),
-    }).passthrough(),
+    data: z.object().passthrough(),
     metadata: z.object().optional(),
 });
 
-export default class Todo extends Document {
+export default class Document extends BaseDocument {
     constructor(options = {}) {
         // Set schema before calling super
         options.schema = options.schema || DOCUMENT_SCHEMA_NAME;
         options.schemaVersion = options.schemaVersion || DOCUMENT_SCHEMA_VERSION;
 
-        // Inject Todo-specific index options BEFORE super()
+        // Inject Document-specific index options BEFORE super() so checksum fields are correct
         options.indexOptions = {
             ...(options.indexOptions || {}),
-            ftsSearchFields: ['data.title', 'data.description'],
-            vectorEmbeddingFields: ['data.title', 'data.description'],
-            checksumFields: ['data.title', 'data.description'],
+            ftsSearchFields: ['data'],
+            vectorEmbeddingFields: ['data'],
+            checksumFields: ['data'],
         };
 
         super(options);
     }
 
-    /**
-     * Create a Todo from minimal data
-     * @param {Object} data - Todo data
-     * @returns {Todo} New Todo instance
-     */
     static fromData(data) {
         data.schema = DOCUMENT_SCHEMA_NAME;
-        return new Todo(data);
+        return new Document(data);
     }
 
     static get dataSchema() {
@@ -50,23 +40,30 @@ export default class Todo extends Document {
     }
 
     static get schema() {
-        return baseDocumentSchema;
+        return documentSchema;
     }
 
     static get jsonSchema() {
         return {
             schema: DOCUMENT_SCHEMA_NAME,
-            data: {
-                title: 'string',
-            },
+            data: {},
         };
     }
 
+    validate() {
+        super.validate();
+    }
+
     static validate(document) {
-        return baseDocumentSchema.parse(document);
+        return documentSchema.parse(document);
+    }
+
+    validateData() {
+        super.validateData();
     }
 
     static validateData(documentData) {
         return documentDataSchema.parse(documentData);
     }
+
 }
