+++ src/schemas/data/Note.js
@@ -3,45 +3,53 @@
 import Document, { documentSchema } from '../BaseDocument.js';
 import { z } from 'zod';
 
-const DOCUMENT_SCHEMA_NAME = 'data/abstraction/tab';
+const DOCUMENT_SCHEMA_NAME = 'data/note';
 const DOCUMENT_SCHEMA_VERSION = '2.0';
 
 const documentDataSchema = z.object({
     schema: z.string(),
     schemaVersion: z.string().optional(),
     data: z.object({
-        url: z.string().url(),
         title: z.string().optional(),
+        content: z.string(),
     }).passthrough(),
     metadata: z.object().optional(),
 });
 
-export default class Tab extends Document {
+export default class Note extends Document {
     constructor(options = {}) {
-        // Set schema defaults before calling super
+        // Set schema before calling super
         options.schema = options.schema || DOCUMENT_SCHEMA_NAME;
         options.schemaVersion = options.schemaVersion || DOCUMENT_SCHEMA_VERSION;
 
-        // Inject Tab-specific index options BEFORE super() so that BaseDocument
-        // computes checksums and other derived values using the correct fields.
+        // Inject Note-specific index options BEFORE super()
         options.indexOptions = {
             ...(options.indexOptions || {}),
-            ftsSearchFields: ['data.title', 'data.url'],
-            vectorEmbeddingFields: ['data.title', 'data.url'],
-            checksumFields: ['data.url'],
+            ftsSearchFields: ['data.title', 'data.content'],
+            vectorEmbeddingFields: ['data.title', 'data.content'],
+            checksumFields: ['data.title', 'data.content'],
         };
 
         super(options);
+
+        if (!this.data.title) {
+            // If Note has no title, we'll use YYYYMMDD, subject to change
+            const now = new Date();
+            const year = now.getFullYear();
+            const month = (now.getMonth() + 1).toString().padStart(2, '0');
+            const day = now.getDate().toString().padStart(2, '0');
+            this.data.title = `${year}${month}${day}`;
+        }
     }
 
     /**
-     * Create a Tab from minimal data
-     * @param {Object} data - Tab data
-     * @returns {Tab} New Tab instance
+     * Create a Note from minimal data
+     * @param {Object} data - Note data
+     * @returns {Note} New Note instance
      */
     static fromData(data) {
         data.schema = DOCUMENT_SCHEMA_NAME;
-        return new Tab(data);
+        return new Note(data);
     }
 
     static get dataSchema() {
@@ -56,22 +64,14 @@ export default class Tab extends Document {
         return {
             schema: DOCUMENT_SCHEMA_NAME,
             data: {
-                url: 'string',
                 title: 'string',
+                content: 'string',
             },
         };
     }
 
-    validate() {
-        super.validate();
-    }
-
     static validate(document) {
         return documentSchema.parse(document);
-    }
-
-    validateData() {
-        super.validateData();
     }
 
     static validateData(documentData) {
