+++ src/schemas/data/Todo.js
@@ -1,38 +1,48 @@
 'use strict';
 
-import BaseDocument, { documentSchema } from '../BaseDocument.js';
+import Document, { documentSchema as baseDocumentSchema } from '../BaseDocument.js';
 import { z } from 'zod';
 
-const DOCUMENT_SCHEMA_NAME = 'data/abstraction/document';
+const DOCUMENT_SCHEMA_NAME = 'data/todo';
 const DOCUMENT_SCHEMA_VERSION = '2.0';
 
 const documentDataSchema = z.object({
     schema: z.string(),
     schemaVersion: z.string().optional(),
-    data: z.object().passthrough(),
+    data: z.object({
+        title: z.string(),
+        description: z.string().optional(),
+        completed: z.boolean().optional(),
+        dueDate: z.string().datetime().optional(),
+    }).passthrough(),
     metadata: z.object().optional(),
 });
 
-export default class Document extends BaseDocument {
+export default class Todo extends Document {
     constructor(options = {}) {
         // Set schema before calling super
         options.schema = options.schema || DOCUMENT_SCHEMA_NAME;
         options.schemaVersion = options.schemaVersion || DOCUMENT_SCHEMA_VERSION;
 
-        // Inject Document-specific index options BEFORE super() so checksum fields are correct
+        // Inject Todo-specific index options BEFORE super()
         options.indexOptions = {
             ...(options.indexOptions || {}),
-            ftsSearchFields: ['data'],
-            vectorEmbeddingFields: ['data'],
-            checksumFields: ['data'],
+            ftsSearchFields: ['data.title', 'data.description'],
+            vectorEmbeddingFields: ['data.title', 'data.description'],
+            checksumFields: ['data.title', 'data.description'],
         };
 
         super(options);
     }
 
+    /**
+     * Create a Todo from minimal data
+     * @param {Object} data - Todo data
+     * @returns {Todo} New Todo instance
+     */
     static fromData(data) {
         data.schema = DOCUMENT_SCHEMA_NAME;
-        return new Document(data);
+        return new Todo(data);
     }
 
     static get dataSchema() {
@@ -40,30 +50,23 @@ export default class Document extends BaseDocument {
     }
 
     static get schema() {
-        return documentSchema;
+        return baseDocumentSchema;
     }
 
     static get jsonSchema() {
         return {
             schema: DOCUMENT_SCHEMA_NAME,
-            data: {},
+            data: {
+                title: 'string',
+            },
         };
     }
 
-    validate() {
-        super.validate();
-    }
-
     static validate(document) {
-        return documentSchema.parse(document);
-    }
-
-    validateData() {
-        super.validateData();
+        return baseDocumentSchema.parse(document);
     }
 
     static validateData(documentData) {
         return documentDataSchema.parse(documentData);
     }
-
 }
