+++ tests/tree.test.js
@@ -16,12 +16,12 @@ describe('SynapsD tree memberships', () => {
     describe('Context tree semantics', () => {
         it('treats "/" as all real memberships in a context tree', async () => {
             const projectId = await db.put({
-                schema: 'data/abstraction/note',
+                schema: 'data/note',
                 data: { title: 'Project Note', content: 'ships in projects only' },
             }, { tree: 'projects', path: '/work/project-a' });
 
             const stagedId = await db.put({
-                schema: 'data/abstraction/note',
+                schema: 'data/note',
                 data: { title: 'Staged Note', content: 'sits only in incoming' },
             }, { tree: 'incoming', path: '/mail/inbox' });
 
@@ -36,7 +36,7 @@ describe('SynapsD tree memberships', () => {
 
         it('refuses unlinking the synthetic context root', async () => {
             const id = await db.put({
-                schema: 'data/abstraction/note',
+                schema: 'data/note',
                 data: { title: 'Protected Root', content: 'context root is a selector, not a real layer' },
             }, { tree: 'projects', path: '/ops/release' });
 
@@ -49,12 +49,12 @@ describe('SynapsD tree memberships', () => {
     describe('Directory tree semantics', () => {
         it('keeps directory "/" literal while still exposing tree memberships', async () => {
             const rootId = await db.put({
-                schema: 'data/abstraction/note',
+                schema: 'data/note',
                 data: { title: 'Root Folder Doc', content: 'lives directly in /' },
             }, { tree: 'incoming', path: '/' });
 
             const nestedId = await db.put({
-                schema: 'data/abstraction/note',
+                schema: 'data/note',
                 data: { title: 'Nested Folder Doc', content: 'lives under /email/inbox' },
             }, { tree: 'incoming', path: '/email/inbox' });
 
@@ -69,7 +69,7 @@ describe('SynapsD tree memberships', () => {
 
         it('removes tree membership after the last directory unlink', async () => {
             const id = await db.put({
-                schema: 'data/abstraction/note',
+                schema: 'data/note',
                 data: { title: 'Multi Path Stage', content: 'linked into two staging folders' },
             }, { tree: 'incoming', path: ['/mail/inbox', '/mail/review'] });
 
@@ -89,7 +89,7 @@ describe('SynapsD tree memberships', () => {
     describe('Incoming promotion flow', () => {
         it('promotes a staged document into a user tree without keeping stale incoming paths', async () => {
             const id = await db.put({
-                schema: 'data/abstraction/note',
+                schema: 'data/note',
                 data: { title: 'Promoted Note', content: 'keep this one' },
             }, { tree: 'incoming', path: '/email/account/inbox' });
